moved filters into file; fixed test
