--- src/strategies/filters/equals.ts
@@ -1,5 +1,0 @@
-import { FilterFn } from '../../types'
-export const equals: FilterFn = (name, type, list) => ({
-  [`${name}`]: type,
-  [`${name}_not`]: type,
-})
